[opentitanlib] Cleanup unused OTP features
OTP support in opentitantool is moving away from end to end vmem
generation. Instead, opentitantool will serve as a front end utility for
generating some software values (such as for alert_handler
configuration) and for providing an way to specifying OTP values using
named constants.
