+++ sw/host/opentitanlib/BUILD
@@ -58,10 +58,7 @@ rust_library(
         "src/otp/alert_handler_regs.rs",
         "src/otp/lc_state.rs",
         "src/otp/mod.rs",
-        "src/otp/otp.rs",
         "src/otp/otp_img.rs",
-        "src/otp/otp_mmap.rs",
-        "src/otp/vmem_serialize.rs",
         "src/proxy/errors.rs",
         "src/proxy/handler.rs",
         "src/proxy/mod.rs",
